+++ recipes/install_server.rb
@@ -16,6 +16,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+node.default['splunk']['service_name'] = 'Splunkd' if systemd?
+node.default['splunk']['startup_script'] = '/etc/systemd/system/Splunkd.service' if systemd?
 
 splunk_installer 'splunk' do
   if upgrade_enabled?