+++ recipes/setup_shclustering.rb
@@ -51,7 +51,7 @@ if node['splunk']['shclustering']['mode'] == 'deployer'
       shcluster_secret: node.run_state['splunk_secret']
     )
     sensitive true unless Chef::Log.debug?
-    notifies :restart, 'service[splunk]', :immediately
+    notifies :restart, 'service[splunk]', :immediately unless disabled?
   end
 end
 
@@ -104,7 +104,7 @@ execute 'initialize search head cluster member' do
     "-conf_deploy_fetch_url #{node['splunk']['shclustering']['deployer_url']} " \
     "-secret #{node.run_state['splunk_secret']} " \
     "-shcluster_label #{node['splunk']['shclustering']['label']}"
-  notifies :restart, 'service[splunk]', :immediately
+  notifies :restart, 'service[splunk]', :immediately unless disabled?
   only_if { init_shcluster_member? }
 end
 
@@ -121,7 +121,7 @@ if ok_to_bootstrap_captain?
     sensitive true unless Chef::Log.debug?
     command "#{splunk_cmd} bootstrap shcluster-captain -auth '#{node.run_state['splunk_auth_info']}' " \
       "-servers_list \"#{shcluster_servers_list.join(',')}\""
-    notifies :restart, 'service[splunk]', :immediately
+    notifies :restart, 'service[splunk]', :immediately unless disabled?
   end
 elsif ok_to_add_member?
   captain_mgmt_uri = "https://#{shcluster_captain}:8089"
@@ -129,7 +129,7 @@ elsif ok_to_add_member?
   execute 'add member to search head cluster' do
     sensitive true unless Chef::Log.debug?
     command "#{splunk_cmd} add shcluster-member -current_member_uri #{captain_mgmt_uri} -auth '#{node.run_state['splunk_auth_info']}'"
-    notifies :restart, 'service[splunk]'
+    notifies :restart, 'service[splunk]' unless disabled?
   end
 end
 
@@ -158,6 +158,6 @@ shpeer_integration_command += ' -site site0' if cluster_master['num_sites'] > 1
 execute 'search head cluster integration with indexer cluster' do
   sensitive true unless Chef::Log.debug?
   command shpeer_integration_command
-  notifies :restart, 'service[splunk]'
+  notifies :restart, 'service[splunk]' unless disabled?
   not_if { search_heads_peered? }
 end