+++ spec/recipes/install_forwarder_spec.rb
@@ -22,26 +22,35 @@ describe 'chef-splunk::install_forwarder' do
     context "#{platform} family" do
       let(:url) { platform_under_test[:url] }
 
-      let(:chef_run) do
-        ChefSpec::ServerRunner.new(platform_under_test[:runner]) do |node|
-          node.force_default['splunk']['forwarder']['version'] = '8.0.1'
-          node.force_default['splunk']['accept_license'] = true
-        end
-      end
-
       context 'url value exists' do
+        let(:chef_run) do
+          ChefSpec::ServerRunner.new(platform_under_test[:runner]) do |node, server|
+            create_data_bag_item(server, 'vault', 'splunk__default')
+            node.force_default['splunk']['forwarder']['version'] = '8.0.1'
+            node.force_default['splunk']['accept_license'] = true
+            node.force_default['chef-vault']['databag_fallback'] = true
+            node.force_default['splunk']['forwarder']['url'] = url
+          end.converge(described_recipe)
+        end
+
         it 'install splunk forwarder from package downloaded from URL' do
-          chef_run.node.force_default['splunk']['forwarder']['url'] = url
-          chef_run.converge(described_recipe)
           expect(chef_run).to run_splunk_installer('splunkforwarder')
             .with(url: url, package_name: 'splunkforwarder')
         end
       end
 
       context 'url attribute is empty' do
+        let(:chef_run) do
+          ChefSpec::ServerRunner.new(platform_under_test[:runner]) do |node, server|
+            create_data_bag_item(server, 'vault', 'splunk__default')
+            node.force_default['splunk']['forwarder']['version'] = '8.0.1'
+            node.force_default['splunk']['accept_license'] = true
+            node.force_default['chef-vault']['databag_fallback'] = true
+            node.force_default['splunk']['forwarder']['url'] = ''
+          end.converge(described_recipe)
+        end
+
         it 'should install splunk forwarder from local repo' do
-          chef_run.node.force_default['splunk']['forwarder']['url'] = ''
-          chef_run.converge(described_recipe)
           expect(chef_run).to run_splunk_installer('splunkforwarder')
             .with(url: '', package_name: 'splunkforwarder')
         end